Fixing timing to comply with international rules, code format

+++ src/co/zer0/raspberrypi/rpimorsecode/TimingScheme.java
@@ -9,40 +9,38 @@ package co.zer0.raspberrypi.rpimorsecode;
  *
  * @author klockhart
  */
-
 //@Singleton
 public class TimingScheme {
-    
+
     //Instance variable
     private static TimingScheme instance;
-    
+
     //Default the unit time to 250ms
     public static final long DEFAULT_UNIT_TIME = 250;
-    
+
     //The global unit time is the length of one unit in uS
     private static long unitTime;
-    
+
     private TimingScheme() {
-        
+
         unitTime = DEFAULT_UNIT_TIME;
     }
-    
-    public static TimingScheme getInstance(){
-        if (instance == null)
-        {
+
+    public static TimingScheme getInstance() {
+        if (instance == null) {
             instance = new TimingScheme();
         }
-        
+
         return instance;
     }
-   
+
     //Get the current global unit time
-    public double getGlobalTimingUnit(){
+    public long getGlobalTimingUnit() {
         return unitTime;
     }
-    
+
     //Set the global unit time
-    public void setGlobalTimingUnit(long newUnitTime){
+    public void setGlobalTimingUnit(long newUnitTime) {
         unitTime = newUnitTime;
     }
 }
